+++ superwall/src/main/java/com/superwall/sdk/models/config/FeatureFlags.kt
@@ -11,6 +11,7 @@ data class RawFeatureFlag(
 
 @Serializable
 data class FeatureFlags(
+    @SerialName("enable_config_refresh") var enableConfigRefresh: Boolean = false,
     @SerialName("enable_session_events") var enableSessionEvents: Boolean,
     @SerialName("enable_postback") var enablePostback: Boolean,
     @SerialName("enable_userid_seed") var enableUserIdSeed: Boolean,