+++ superwall/src/main/java/com/superwall/sdk/paywall/presentation/internal/operators/GetPresenter.kt
@@ -2,6 +2,8 @@ package com.superwall.sdk.paywall.presentation.internal.operators
 
 import android.app.Activity
 import com.superwall.sdk.Superwall
+import com.superwall.sdk.analytics.internal.track
+import com.superwall.sdk.analytics.internal.trackable.InternalSuperwallEvent
 import com.superwall.sdk.delegate.SubscriptionStatus
 import com.superwall.sdk.logger.LogLevel
 import com.superwall.sdk.logger.LogScope
@@ -12,6 +14,7 @@ import com.superwall.sdk.paywall.presentation.internal.InternalPresentationLogic
 import com.superwall.sdk.paywall.presentation.internal.PaywallPresentationRequestStatusReason
 import com.superwall.sdk.paywall.presentation.internal.PresentationRequest
 import com.superwall.sdk.paywall.presentation.internal.PresentationRequestType
+import com.superwall.sdk.paywall.presentation.internal.request.PresentationInfo
 import com.superwall.sdk.paywall.presentation.internal.state.PaywallSkippedReason
 import com.superwall.sdk.paywall.presentation.internal.state.PaywallState
 import com.superwall.sdk.paywall.presentation.rule_logic.RuleEvaluationOutcome
@@ -50,11 +53,10 @@ suspend fun Superwall.getPresenterIfNecessary(
     when (request.flags.type) {
         is PresentationRequestType.GetPaywall -> {
             val sessionId =
-                activateSession(
+                attemptTriggerFire(
                     request = request,
                     triggerResult = rulesOutcome.triggerResult,
                 )
-            paywallView.paywall.triggerSessionId = sessionId
             return null
         }
 
@@ -66,11 +68,10 @@ suspend fun Superwall.getPresenterIfNecessary(
     }
 
     val sessionId =
-        activateSession(
+        attemptTriggerFire(
             request = request,
             triggerResult = rulesOutcome.triggerResult,
         )
-    paywallView.paywall.triggerSessionId = sessionId
 
     val currentActivity = dependencyContainer.activityProvider?.getCurrentActivity()
 
@@ -94,13 +95,34 @@ suspend fun Superwall.getPresenterIfNecessary(
     return currentActivity
 }
 
-private suspend fun Superwall.activateSession(
+suspend fun Superwall.attemptTriggerFire(
     request: PresentationRequest,
     triggerResult: InternalTriggerResult,
-): String? {
+) {
+    if (request.flags.type == PresentationRequestType.GetImplicitPresentationResult ||
+        request.flags.type == PresentationRequestType.GetPresentationResult
+    ) {
+        return
+    }
     val sessionEventsManager = dependencyContainer.sessionEventsManager
-    return sessionEventsManager?.triggerSession?.activateSession(
-        request.presentationInfo,
-        triggerResult,
-    )
+    // If eventName is null, the paywall is being presented by identifier, which is what the debugger uses and that's not supported.
+    val eventName = request.presentationInfo.eventName ?: return
+
+    when (val req = request.presentationInfo) {
+        is PresentationInfo.ExplicitTrigger, is PresentationInfo.ImplicitTrigger -> {
+            when (triggerResult) {
+                is InternalTriggerResult.Error, is InternalTriggerResult.EventNotFound ->
+                    return
+                else -> {} // No-op
+            }
+        }
+        is PresentationInfo.FromIdentifier -> {} // No-op
+    }
+    val trackedEvent =
+        InternalSuperwallEvent.TriggerFire(
+            triggerResult = triggerResult,
+            triggerName = eventName,
+            sessionEventsManager = sessionEventsManager,
+        )
+    track(trackedEvent)
 }