+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall/Superwall-Android/releases) on GitHub.
 
+
+## 1.2.1
+
+### Enhancements
+
+- Adds the ability for the SDK to refresh the Superwall configuration every session start, subject to a feature flag.
+- Tracks a `config_refresh` Superwall event when the configuration is refreshed.
+- SW-2891: Adds` abandoned_product_id` to a `transaction_abandon` event to use in audience filters. You can use this to show a paywall if a user abandons the transaction for a specific product.
+
 ## 1.2.0
 
 ### Enhancements